sundry

+++ setup-shared.sh
@@ -4,7 +4,7 @@
 ln -sf ~/config/bin ~
 ln -sf ~/config/mac-bashrc ~/.bashrc
 ln -sf ~/config/init.el ~/.emacs
-ln -sf ~/config/emacs.d ~/.emacs.d
+test -e ~/.emacs.d || ln -sf ~/config/emacs.d ~/.emacs.d
 ln -sf ~/config/gitconfig ~/.gitconfig
 ln -sf ~/config/tmux ~/.tmux.conf
 ln -sf ~/config/quiltrc  ~/.quiltrc
